refactor: barbershop components
