+++ app/barbershops/[id]/_action/SaveBooking.ts
@@ -3,7 +3,6 @@
 import { db } from "@/app/_lib/prisma";
 import { revalidatePath } from "next/cache";
 
-//Aula 3.2 Backend da reserva
 interface SaveBookingParams {
   barbershopId: string;
   serviceId: string;
@@ -21,7 +20,6 @@ export const saveBooking = async (params: SaveBookingParams) => {
       }
    })
 
-   //Serve para atualizar a rota, evitando que não apareça a reserva para o usuário
    revalidatePath("/bookings");
    revalidatePath("/");
 };
